Update patch set 1
Patch Set 1:

I would suggest to place this charm at github and while running the script download it form github. Its not a good idea to include the static charm in the code. As juju changes and in future it might fail and we can not change the OPNFV code.

Patch-set: 1